requirements.txt
amqp==5.3.1
annotated-types==0.7.0
anthropic==0.69.0
anyio==4.11.0
asgiref==3.9.1
async-timeout==5.0.1
attrs==25.3.0
Authlib==1.6.5
billiard==4.2.2
celery==5.3.6
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.3.0
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cloudinary==1.44.1
colorama==0.4.6
cryptography==46.0.2
cyclopts==3.24.0
distro==1.9.0
dj-rest-auth==7.0.1
Django==5.0.2
django-allauth==65.11.2
django-cloudinary-storage==0.3.0
django-cors-headers==4.3.1
django-filter==24.1
django-redis==6.0.0
djangorestframework==3.15.0
djangorestframework_simplejwt==5.5.1
dnspython==2.8.0
docstring_parser==0.17.0
docutils==0.22.2
drf-spectacular==0.27.1
email-validator==2.3.0
exceptiongroup==1.3.0
fastmcp==2.12.5
gunicorn==21.2.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.1
idna==3.10
inflection==0.5.1
isodate==0.7.2
jiter==0.11.0
jsonschema==4.25.1
jsonschema-path==0.3.4
jsonschema-specifications==2025.9.1
python-decouple>=3.8
